neue methoden added

+++ course223/src/main/java/ch/noseryoung/course223/domainModels/Function.java
@@ -3,5 +3,5 @@ package ch.noseryoung.course223.domainModels;
 @FunctionalInterface
 public interface Function{
 
-    public int calculate(int a, int b);
+    public double calculate(double a, double b);
 }
